Routed received commands to the CLI

+++ pico/emulator-w/src/cli.c
@@ -15,58 +15,74 @@
 
 typedef void (*handler)(uint32_t, uint32_t);
 
-void query();
-void reboot();
-void help();
-
+void help(txrx f, void *context);
+void query(txrx f, void *context);
 void on_card_command(char *cmd, handler fn);
 void on_door_open();
 void on_door_close();
 void on_press_button();
 void on_release_button();
 void write(uint32_t, uint32_t);
+void reboot();
 
-/* Executes a command.
+void serial(void *context, const char *msg) {
+    puts(msg);
+}
+
+/* Executes a command from the TTY terminal.
  *
  */
 void exec(char *cmd) {
     int N = strlen(cmd);
 
     if (N > 0) {
-        if (strncasecmp(cmd, "query", 5) == 0) {
-            query();
-        } else if (strncasecmp(cmd, "open", 4) == 0) {
-            on_door_open();
-        } else if (strncasecmp(cmd, "close", 5) == 0) {
-            on_door_close();
-        } else if (strncasecmp(cmd, "press", 5) == 0) {
-            on_press_button();
-        } else if (strncasecmp(cmd, "release", 7) == 0) {
-            on_release_button();
-        } else if (strncasecmp(cmd, "reboot", 6) == 0) {
-            reboot();
-        } else if (strncasecmp(cmd, "cls", 5) == 0) {
+        if (strncasecmp(cmd, "cls", 3) == 0) {
             cls();
             return; // don't invoke clearline because that puts a >> prompt in the wrong place
-        } else if ((cmd[0] == 't') || (cmd[0] == 'T')) {
-            sys_settime(&cmd[1]);
-        } else if ((cmd[0] == 'w') || (cmd[0] == 'W')) {
-            on_card_command(&cmd[1], write);
         } else {
-            help();
+            execw(cmd, serial, NULL);
         }
     }
 
     clearline();
 }
 
+void execw(char *cmd, txrx f, void *data) {
+    if (f != NULL) {
+        int N = strlen(cmd);
+
+        if (N > 0) {
+            if (strncasecmp(cmd, "query", 5) == 0) {
+                query(f, data);
+            } else if (strncasecmp(cmd, "open", 4) == 0) {
+                on_door_open();
+            } else if (strncasecmp(cmd, "close", 5) == 0) {
+                on_door_close();
+            } else if (strncasecmp(cmd, "press", 5) == 0) {
+                on_press_button();
+            } else if (strncasecmp(cmd, "release", 7) == 0) {
+                on_release_button();
+            } else if (strncasecmp(cmd, "reboot", 6) == 0) {
+                reboot();
+            } else if ((cmd[0] == 't') || (cmd[0] == 'T')) {
+                sys_settime(&cmd[1]);
+            } else if ((cmd[0] == 'w') || (cmd[0] == 'W')) {
+                on_card_command(&cmd[1], write);
+            } else {
+                help(f, data);
+            }
+        }
+    }
+}
+
 /* Displays the last read/write card, if any.
  *
  */
-void query() {
+void query(txrx f, void *context) {
     char s[64];
+
     cardf(&last_card, s, sizeof(s));
-    puts(s);
+    f(context, s);
 }
 
 /* Write card command.
@@ -82,9 +98,7 @@ void write(uint32_t facility_code, uint32_t card) {
 /* Goes into a tight loop until the watchdog resets the processor.
  *
  */
-void reboot() {
-    tcpd_terminate();
-
+void reboot(txrx f, void *context) {
     while (true) {
         buzzer_beep(1);
 
@@ -105,29 +119,20 @@ void reboot() {
 /* Displays a list of the supported commands.
  *
  */
-void help() {
-    static uint8_t x = 0x00;
-
-    TPIC_set(RED_LED, (x & 0x01) == 0x01);
-    TPIC_set(ORANGE_LED, (x & 0x02) == 0x02);
-    TPIC_set(YELLOW_LED, (x & 0x04) == 0x04);
-    TPIC_set(GREEN_LED, (x & 0x08) == 0x08);
-
-    x++;
-
-    tx("-----");
-    tx("Commands:");
-    tx("TYYYY-MM-DD HH:mm:ss  Set date/time");
-    tx("Wnnnnnn               Write card to Wiegand-26 interface");
-    tx("QUERY                 Display last card read/write");
-    tx("OPEN                  Opens door contact relay");
-    tx("CLOSE                 Closes door contact relay");
-    tx("PRESS                 Press pushbutton");
-    tx("RELEASE               Release pushbutton");
-    tx("CLS                   Resets the terminal");
-    tx("REBOOT                Reboot");
-    tx("?                     Display list of commands");
-    tx("-----");
+void help(txrx f, void *context) {
+    f(context, "-----");
+    f(context, "Commands:");
+    f(context, "TYYYY-MM-DD HH:mm:ss  Set date/time");
+    f(context, "Wnnnnnn               Write card to Wiegand-26 interface");
+    f(context, "QUERY                 Display last card read/write");
+    f(context, "OPEN                  Opens door contact relay");
+    f(context, "CLOSE                 Closes door contact relay");
+    f(context, "PRESS                 Press pushbutton");
+    f(context, "RELEASE               Release pushbutton");
+    f(context, "CLS                   Resets the terminal");
+    f(context, "REBOOT                Reboot");
+    f(context, "?                     Display list of commands");
+    f(context, "-----");
 }
 
 /* Card command handler.
@@ -167,11 +172,11 @@ void on_card_command(char *cmd, handler fn) {
  *  Opens/closes the door contact emulation relay (in reader mode only).
  *
  */
-void on_door_open() {
+void on_door_open(txrx f, void *context) {
     relay_door_contact(false);
 }
 
-void on_door_close(char *cmd) {
+void on_door_close(txrx f, void *context) {
     relay_door_contact(true);
 }
 
@@ -179,10 +184,10 @@ void on_door_close(char *cmd) {
  *  Opens/closes the pushbutton emulation relay (in reader mode only).
  *
  */
-void on_press_button() {
+void on_press_button(txrx f, void *context) {
     relay_pushbutton(true);
 }
 
-void on_release_button() {
+void on_release_button(txrx f, void *context) {
     relay_pushbutton(false);
 }
